Code review changes have been applied.
