+++ spring-cloud-task-samples/timestamp/pom.xml
@@ -5,7 +5,7 @@
 
 	<modelVersion>4.0.0</modelVersion>
 
-	<groupId>org.springframework.cloud.task.module</groupId>
+	<groupId>org.springframework.cloud</groupId>
 	<artifactId>timestamp-task</artifactId>
 	<packaging>jar</packaging>
 	<name>timestamp-task</name>
@@ -15,11 +15,11 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>1.3.0.BUILD-SNAPSHOT</version>
+		<version>1.3.0.RELEASE</version>
 	</parent>
 
 	<properties>
-		<start-class>demo.TaskApplication</start-class>
+		<start-class>org.springframework.cloud.task.timestamp.TaskApplication</start-class>
 	</properties>
 
 	<dependencies>
@@ -46,10 +46,6 @@
 			<artifactId>hsqldb</artifactId>
 		</dependency>
 		<dependency>
-			<groupId>org.hibernate</groupId>
-			<artifactId>hibernate-validator</artifactId>
-		</dependency>
-		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-configuration-processor</artifactId>
 			<optional>true</optional>
@@ -61,9 +57,6 @@
 			<plugin>
 				<groupId>org.springframework.boot</groupId>
 				<artifactId>spring-boot-maven-plugin</artifactId>
-				<configuration>
-					<classifier>exec</classifier>
-				</configuration>
 			</plugin>
 		</plugins>
 	</build>
