+++ spring-cloud-task-samples/timestamp/src/main/java/org/springframework/cloud/task/timestamp/TaskApplication.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package demo;
+package org.springframework.cloud.task.timestamp;
 
 
 import java.text.DateFormat;
