+++ spring-cloud-task-samples/timestamp/src/test/java/org/springframework/cloud/task/timestamp/TimestampTaskPropertiesTests.java
@@ -1,5 +1,6 @@
 /*
  * Copyright 2015 the original author or authors.
+ *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -13,13 +14,11 @@
  * limitations under the License.
  */
 
-package demo;
+package org.springframework.cloud.task.timestamp;
 
 import static org.junit.Assert.assertEquals;
 
 import org.junit.Test;
-
-import org.springframework.beans.factory.BeanCreationException;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.boot.test.EnvironmentTestUtils;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
@@ -30,13 +29,14 @@ import org.springframework.context.annotation.Configuration;
  */
 public class TimestampTaskPropertiesTests {
 
-	@Test(expected = BeanCreationException.class)
+	@Test(expected = IllegalArgumentException.class)
 	public void testEmptyFormat() {
 		AnnotationConfigApplicationContext context = new AnnotationConfigApplicationContext();
 		EnvironmentTestUtils.addEnvironment(context, "format:");
 		context.register(Conf.class);
 		context.refresh();
 		TimestampTaskProperties properties = context.getBean(TimestampTaskProperties.class);
+		properties.getFormat();
 	}
 
 	@Test
